requirements.txt
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
Django==1.11.16
django-cors-headers==2.4.0
django-filter==2.0.0
django-model-utils==2.0.3
djangorestframework==3.9.0
idna==2.7
kiwisolver==1.1.0
Markdown==3.0.1
matplotlib==3.0.3
numpy==1.16.4
Pillow==6.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.7
requests==2.20.1
six==1.12.0
urllib3==1.24.1